+++ src/main/java/mod/azure/doom/entity/projectiles/RocketEntity.java
@@ -272,7 +272,7 @@ public class RocketEntity extends AbstractArrow implements IAnimatable {
 			if (d12 <= 1.0D) {
 				if (entity instanceof LivingEntity) {
 					entity.hurt(DamageSource.playerAttack((Player) this.shooter),
-							DoomConfig.SERVER.rocket_damage.get());
+							DoomConfig.SERVER.rocket_damage.get().floatValue());
 				}
 			}
 		}
