+++ src/main/java/mod/azure/doom/entity/tierboss/IconofsinEntity.java
@@ -303,7 +303,7 @@ public class IconofsinEntity extends DemonEntity implements IAnimatable {
 			double d12 = (double) (Mth.sqrt((float) entity.distanceToSqr(vector3d)) / f2);
 			if (d12 <= 1.0D) {
 				if (entity instanceof LivingEntity) {
-					entity.hurt(DamageSource.indirectMobAttack(this, this.getTarget()), DoomConfig.SERVER.icon_melee_damage.get());
+					entity.hurt(DamageSource.indirectMobAttack(this, this.getTarget()), DoomConfig.SERVER.icon_melee_damage.get().floatValue());
 				}
 			}
 		}
