+++ src/main/java/mod/azure/doom/entity/tierboss/SpiderMastermind2016Entity.java
@@ -36,7 +36,7 @@ public class SpiderMastermind2016Entity extends SpiderMastermindEntity {
 		this.goalSelector.addGoal(4,
 				new RangedStrafeAttackGoal(this,
 						new SpiderMastermind2016Entity.FireballAttack(this).setProjectileOriginOffset(0.8, 0.2, 0.8)
-								.setDamage(DoomConfig.SERVER.spider_mastermind_ranged_damage.get()),
+								.setDamage(DoomConfig.SERVER.spider_mastermind_ranged_damage.get().floatValue()),
 						1.0D, 50, 30, 15, 15F, 1).setMultiShot(5, 1));
 		this.targetSelector.addGoal(1, new HurtByTargetGoal(this).setAlertOthers());
 		this.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(this, Player.class, true));
