+++ src/main/java/mod/azure/doom/entity/tierfodder/GargoyleEntity.java
@@ -115,7 +115,7 @@ public class GargoyleEntity extends DemonEntity implements IAnimatable, Enemy {
 		this.goalSelector.addGoal(4,
 				new RangedStrafeAttackGoal(this,
 						new FireballAttack(this, false).setProjectileOriginOffset(0.8, 0.8, 0.8)
-								.setDamage(DoomConfig.SERVER.gargoyle_ranged_damage.get()).setSound(SoundEvents.BLAZE_SHOOT, 1.0F,
+								.setDamage(DoomConfig.SERVER.gargoyle_ranged_damage.get().floatValue()).setSound(SoundEvents.BLAZE_SHOOT, 1.0F,
 										1.4F + this.getRandom().nextFloat() * 0.35F),
 						1.0D, 50, 30, 15, 15F, 1).setMultiShot(3, 3));
 		this.goalSelector.addGoal(7, new GargoyleEntity.LookAroundGoal(this));
