+++ src/main/java/mod/azure/doom/entity/tierfodder/MechaZombieEntity.java
@@ -106,7 +106,7 @@ public class MechaZombieEntity extends DemonEntity implements IAnimatable {
 		this.goalSelector.addGoal(4,
 				new RangedStrafeAttackGoal(this,
 						new FireballAttack(this, false).setProjectileOriginOffset(0.8, 0.8, 0.8)
-								.setDamage(DoomConfig.SERVER.mechazombie_ranged_damage.get()).setSound(SoundEvents.BLAZE_SHOOT,
+								.setDamage(DoomConfig.SERVER.mechazombie_ranged_damage.get().floatValue()).setSound(SoundEvents.BLAZE_SHOOT,
 										1.0F, 1.4F + this.getRandom().nextFloat() * 0.35F),
 						1.0D, 50, 30, 15, 15F, 1).setMultiShot(4, 3));
 		this.goalSelector.addGoal(4, new DemonAttackGoal(this, 1.0D, false, 2));
