+++ src/main/java/mod/azure/doom/entity/tierheavy/RevenantEntity.java
@@ -93,7 +93,7 @@ public class RevenantEntity extends DemonEntity implements IAnimatable {
 		this.goalSelector.addGoal(4,
 				new RangedStrafeAttackGoal(this,
 						new RevenantEntity.FireballAttack(this).setProjectileOriginOffset(0.8, 0.8, 0.8)
-								.setDamage(DoomConfig.SERVER.revenant_ranged_damage.get()),
+								.setDamage(DoomConfig.SERVER.revenant_ranged_damage.get().floatValue()),
 						1.0D, 50, 30, 15, 15F, 1).setMultiShot(2, 3));
 		this.goalSelector.addGoal(4, new DemonAttackGoal(this, 1.0D, false, 2));
 		this.goalSelector.addGoal(5, new WaterAvoidingRandomStrollGoal(this, 0.8D));
