+++ src/main/java/mod/azure/doom/entity/tiersuperheavy/DoomHunterEntity.java
@@ -274,7 +274,7 @@ public class DoomHunterEntity extends DemonEntity implements IAnimatable {
 				float f = (float) Mth.atan2(livingentity.getZ() - parentEntity.getZ(),
 						livingentity.getX() - parentEntity.getX());
 				RocketMobEntity fireballentity = new RocketMobEntity(world, this.parentEntity, d2, d3, d4,
-						DoomConfig.SERVER.doomhunter_ranged_damage.get());
+						DoomConfig.SERVER.doomhunter_ranged_damage.get().floatValue());
 				if (this.attackTimer == 15) {
 					if (parentEntity.getHealth() < (parentEntity.getMaxHealth() * 0.50)) {
 						for (int l = 0; l < 16; ++l) {
