+++ src/main/java/mod/azure/doom/entity/tiersuperheavy/MarauderEntity.java
@@ -125,7 +125,7 @@ public class MarauderEntity extends DemonEntity implements IAnimatable {
 		this.goalSelector.addGoal(4, new DemonAttackGoal(this, 1.0D, false, 1));
 		this.goalSelector.addGoal(4,
 				new RangedStaticAttackGoal(this, new MarauderEntity.FireballAttack(this)
-						.setProjectileOriginOffset(0.8, 0.8, 0.8).setDamage(DoomConfig.SERVER.shotgun_damage.get()), 60, 20, 30F,
+						.setProjectileOriginOffset(0.8, 0.8, 0.8).setDamage(DoomConfig.SERVER.shotgun_damage.get().floatValue()), 60, 20, 30F,
 						2));
 		this.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(this, Player.class, true));
 		this.targetSelector.addGoal(3, new NearestAttackableTargetGoal<>(this, AbstractVillager.class, false));
